# easy-list/openapi-definition.yaml
openapi: 3.0.1
info:
  title: EasyList
  version: v1
paths:
  /ListItems:
    get:
      tags:
        - ListItems
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListItem'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListItem'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListItem'
    post:
      tags:
        - ListItems
      requestBody:
        content:
          application/json:
            schema:
              type: string
              nullable: true
          text/json:
            schema:
              type: string
              nullable: true
          application/*+json:
            schema:
              type: string
              nullable: true
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ListItem'
            application/json:
              schema:
                $ref: '#/components/schemas/ListItem'
            text/json:
              schema:
                $ref: '#/components/schemas/ListItem'
  '/ListItems/{id}':
    get:
      tags:
        - ListItems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - ListItems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Success
components:
  schemas:
    ListItem:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }